Do not allow redirects and enforce pcloud domain
This is security fix.
Following of redirects sent from malicious user impersonating pcloud servers
can result in downloading of unprotected http resources accessible from the
local network. Depending on the further processing of the content of those
resources they may be leaked to the attacker (e.g., if they are being uploaded
to some location that is under attacker's control).

Thanks to Joey Hess for pointing that out.